+++ cqi_cpp/src/state.cpp
@@ -1,5 +1,5 @@
 #include "../include/state.hpp"
 
-State::State(vector<float>* state) {
+State::State(vector<double>* state) {
     this->state = state;
 }
